feat: Remove Header
Remove header and incorporate everything else into the side bar and other areas
